pyproject.toml
[tool.poetry]
name = "celery-django-case"
version = "0.1.0"
description = "Understanding Celery in a Django app"
authors = ["Rodrigo Farah"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
django = "^5.2"
celery = "^5.5.2"
django-stubs = "^5.2.0"
python-dotenv = "^1.1.0"
webdriver-manager = "^4.0.2"
django-celery-results = "^2.6.0"
django-celery-beat = "^2.8.0"
django-rest-framework = "^0.1.0"
psycopg2-binary = "^2.9.10"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
